Add Data.Map.Ordered.Strict
To facilitate code reuse between `Data.Map.Ordered` and
`Data.Map.Ordered.Strict`, I migrated the existing `Data.Map.Ordered`
module to `Data.Map.Ordered.Internal`.

Fixes #10.
