+++ ordered-containers.cabal
@@ -15,8 +15,8 @@ source-repository head
   location:            https://github.com/dmwit/ordered-containers
 
 library
-  exposed-modules:     Data.Map.Ordered, Data.Set.Ordered
-  other-modules:       Data.Map.Util
+  exposed-modules:     Data.Map.Ordered, Data.Map.Ordered.Strict, Data.Set.Ordered
+  other-modules:       Data.Map.Ordered.Internal, Data.Map.Util
   build-depends:       base >=4.7 && <5, containers >=0.1 && <0.7
   default-language:    Haskell98
   ghc-options:         -fno-warn-tabs
